+++ R/multitables.R
@@ -168,6 +168,14 @@ newMultitable <- function (formula, data, name, ...) {
     }
 
     template <- formulaToQuery(formula, data)
+    if (has.selected.array(template)) {
+        halt("Multitables do not support specifications that use deprecated",
+             " multiple response functions. Do you have `options(crunch.mr.",
+             "selection = \"selected_array\")` set? If so, please change it",
+             " to `options(crunch.mr.selection = \"as_selected\")` and try ",
+             "to create the multitable over again.")
+    }
+    
     if (missing(name)) {
         name <- RHS_string(formula)
     }
@@ -177,6 +185,21 @@ newMultitable <- function (formula, data, name, ...) {
                 body=toJSON(payload))
 
     invisible(Multitable(crGET(u)))
+}
+
+# check if a template or query has a selected_array somewhere in it recursively.
+has.selected.array <- function (query) {
+    clauses <- lapply(query, function (clause) {
+        if (!is.list(clause)) {
+            # there are no more levels to traverse
+            return(FALSE)
+        } else if (clause[["function"]] %||% FALSE == "selected_array") {
+            return(TRUE)
+        } 
+        return(lapply(clause, has.selected.array))
+    })
+    
+    return(any(unlist(clauses)))
 }
 
 #' Import a Multitable