+++ tests/testthat/test-active-filter.R
@@ -208,10 +208,11 @@ with_test_authentication({
     test_that("as.data.frame when filtered", {
         df2 <- as.data.frame(ds2)
         expect_identical(df2$v3, c(9, 11, 13, 15, 17, 19, 21, 23, 25, 27))
-        expect_equivalent(as.data.frame(ds2[,c("v3", "v4")], force=TRUE),
-            df[df$v4 == "C", c("v3", "v4")])
         df3 <- as.data.frame(ds3)
         expect_equivalent(df3$v3, 12:27)
+        skip_locally("Vagrant host doesn't serve files correctly")
+        expect_equivalent(as.data.frame(ds2[,c("v3", "v4")], force=TRUE),
+                          df[df$v4 == "C", c("v3", "v4")])
     })
 
     test_that("filtered cubing", {